feat: Conditionally display set scores on ScoreScreen
This commit introduces a setting to control the visibility of the set scores on the ScoreScreen.

The UI element that displays the number of sets won by each player is now wrapped in a conditional block, `if (currentSettings.showSets)`. This allows the set score display to be hidden based on the user's preferences.